Major refactoring of how triggers work
The central routing with (triggers) has been replaced with a
(deftriggers) macro, which declares a name, signature, help text, list
of triggers, and handler body for a given action in one place.

This will make adding things like !help much easier, as it will be
possible to save the help text for a command and automatically look it
up with !help.